+++ lib/llm/src/discovery/model_manager.rs
@@ -15,7 +15,7 @@ use crate::discovery::runtime_configs::RuntimeConfigs;
 
 use dynamo_runtime::{
     component::{Client, Endpoint, build_transport_type},
-    discovery::{DiscoveryQuery, DiscoverySpec, watch_and_extract_field},
+    discovery::DiscoverySpec,
     prelude::DistributedRuntimeProvider,
     protocols::EndpointId,
 };
@@ -616,7 +616,7 @@ impl ModelManager {
     }
 
     /// Get or create a runtime config watcher for an endpoint.
-    /// Spawns a background task to watch DiscoveryQuery::EndpointModels.
+    /// Spawns a background task to watch for worker config changes.
     /// Returns a shared RuntimeConfigs that KvScheduler can use directly.
     pub async fn get_or_create_runtime_config_watcher(
         &self,
@@ -641,8 +641,7 @@ impl ModelManager {
 
         // Only spawn watcher if we were the one who inserted
         if is_new {
-            self.spawn_runtime_config_watcher(endpoint, result.clone())
-                .await?;
+            result.start_watcher(endpoint).await?;
         }
 
         Ok(result)
@@ -658,81 +657,6 @@ impl ModelManager {
         let inner = self.runtime_configs.get(endpoint_id)?;
         let config_ref = inner.configs.get(&worker_id)?;
         config_ref.as_ref()?.disaggregated_endpoint.clone()
-    }
-
-    /// Spawn background task to watch runtime configs via discovery.
-    /// Does not block - consumers should use `subscribe().wait_for_some()` if they need workers.
-    async fn spawn_runtime_config_watcher(
-        &self,
-        endpoint: &Endpoint,
-        inner: Arc<RuntimeConfigs>,
-    ) -> anyhow::Result<()> {
-        let component = endpoint.component();
-        let cancellation_token = component.drt().primary_token();
-
-        // Set up discovery watch for EndpointModels
-        let discovery = component.drt().discovery();
-        let endpoint_id = endpoint.id();
-        let discovery_key = DiscoveryQuery::EndpointModels {
-            namespace: endpoint_id.namespace.clone(),
-            component: endpoint_id.component.clone(),
-            endpoint: endpoint_id.name.clone(),
-        };
-        let discovery_stream = discovery
-            .list_and_watch(discovery_key.clone(), Some(cancellation_token.clone()))
-            .await?;
-
-        // Extract runtime_config from ModelDeploymentCard
-        let mut runtime_configs_rx =
-            watch_and_extract_field(discovery_stream, |card: ModelDeploymentCard| {
-                card.runtime_config
-            });
-
-        // Also watch instance IDs
-        let client = endpoint.client().await?;
-        let mut instance_ids_rx = client.instance_avail_watcher();
-
-        // Spawn background task to watch for config changes
-        // Note: We don't block here - consumers should wait on notify for configs they need
-        let cancel_token = cancellation_token.clone();
-        tokio::spawn(async move {
-            tracing::trace!("ModelManager runtime config watcher started");
-            loop {
-                // Wait for either instances or configs to change
-                tokio::select! {
-                    _ = cancel_token.cancelled() => {
-                        tracing::trace!("ModelManager runtime config watcher shutting down");
-                        break;
-                    }
-                    result = instance_ids_rx.changed() => {
-                        if result.is_err() {
-                            tracing::warn!("instance IDs watch sender shutdown in ModelManager");
-                            break;
-                        }
-                    }
-                    result = runtime_configs_rx.changed() => {
-                        if result.is_err() {
-                            tracing::warn!("runtime configs watch sender shutdown in ModelManager");
-                            break;
-                        }
-                    }
-                }
-
-                // Get the latest values from both channels
-                let new_instance_ids = instance_ids_rx.borrow_and_update().clone();
-                let new_configs = runtime_configs_rx.borrow_and_update().clone();
-
-                inner.update(&new_instance_ids, &new_configs);
-
-                tracing::trace!(
-                    "ModelManager: Updated runtime_configs with {} workers",
-                    inner.configs.len()
-                );
-            }
-            tracing::trace!("ModelManager runtime config watcher shutting down");
-        });
-
-        Ok(())
     }
 
     /// Lists all models that have worker monitors (and thus busy thresholds) configured.
